add fragment navigation
