--- app/src/main/java/com/shchurovsi/todolistplanner/presentation/fragments/HomeScreenFragment.kt
@@ -1,25 +1,0 @@
-package com.shchurovsi.todolistplanner.presentation.fragments
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.shchurovsi.todolistplanner.R
-
-
-class HomeScreenFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_screen, container, false)
-    }
-
-    companion object {
-        @JvmStatic
-        fun newInstance() = HomeScreenFragment()
-    }
-}
